Add Namespace and Command classes
The Namespaces are wrappers around dicts, Commands are wrappers around
callables. Both exist in an unbound and a bound form, the unbound form
as the additional method `bind` to bind the command / all the commands
in a namespace to an arbitrary object.

This replaces the `build_namespace_from_object` method with
`Namespace.from_object`.