Add support for nested namespaces
Nested namespaces can be used to describe command namespaces inside
attributes of the top-level object. This can be used if sub-modules of
drivers are separated through properties.

When the sub-namespaces are parsed first, they can be passed into the
parent namespace. Commands will be in the parent-namespace with a prefix
for the child namespace